+++ UnBBayes/src/unbbayes/datamining/classifiers/DistributionClassifier.java
@@ -1,8 +1,9 @@
 package unbbayes.datamining.classifiers;
 
-import java.util.*;
+import java.util.ResourceBundle;
 
-import unbbayes.datamining.datamanipulation.*;
+import unbbayes.datamining.datamanipulation.Instance;
+import unbbayes.datamining.datamanipulation.Utils;
 
 /**
  *  Abstract Bayesian Classifier. All schemes that works with bayesian classifiers extends this class.
@@ -51,14 +52,14 @@ public abstract class DistributionClassifier extends Classifier
    	* Instance.missingValue() if no prediction is made
    	* @exception Exception if an error occurred during the prediction
    	*/
-  	public int classifyInstance(Instance instance) throws Exception
+  	public short classifyInstance(Instance instance) throws Exception
 	{   float[] dist = distributionForInstance(instance);
             if (dist == null)
             {	throw new Exception(resource.getString("nullPrediction"));
             }
             else
             {   switch (classificationType)
-                {   case NORMAL_CLASSIFICATION :              return Utils.maxIndex(dist);
+                {   case NORMAL_CLASSIFICATION :              return (byte)Utils.maxIndex(dist);
                     case RELATIVE_FREQUENCY_CLASSIFICATION :  int i,maxIndex = -1;
                                                               float max = Float.MIN_VALUE;
                                                               float local;
@@ -69,7 +70,7 @@ public abstract class DistributionClassifier extends Classifier
                                                                       maxIndex = i;
                                                                   }
                                                               }
-                                                              return maxIndex;
+                                                              return (byte)maxIndex;
                     case ABSOLUTE_FREQUENCY_CLASSIFICATION :  int j;
                                                               for (j=0; j<classValues.length; j++)
                                                               {   int actualValue = classValues[j];
@@ -77,7 +78,7 @@ public abstract class DistributionClassifier extends Classifier
                                                                   if (dist[actualValue] >= probabilities[j])
                                                                   {   //System.out.println("Predic "+dist[actualValue]);
                                                                       //System.out.println("User "+probabilities[j]);
-                                                                  return actualValue;
+                                                                  return (byte)actualValue;
                                                                   }
                                                               }
                                                               return Instance.MISSING_VALUE;
@@ -86,7 +87,7 @@ public abstract class DistributionClassifier extends Classifier
             }
       }
 
-	public int classifyInstance(float[] dist) throws Exception
+	public byte classifyInstance(float[] dist) throws Exception
 	{   
 		if (dist == null)
 		{	
@@ -96,7 +97,7 @@ public abstract class DistributionClassifier extends Classifier
 		{   
 			switch (classificationType)
 			{   
-				case NORMAL_CLASSIFICATION :              	return Utils.maxIndex(dist);
+				case NORMAL_CLASSIFICATION :              	return (byte)Utils.maxIndex(dist);
 				case RELATIVE_FREQUENCY_CLASSIFICATION :  	int i,maxIndex = -1;
 															float max = Float.MIN_VALUE;
 															float local;
@@ -109,14 +110,14 @@ public abstract class DistributionClassifier extends Classifier
 																	maxIndex = i;
 																}
 															}
-															return maxIndex;
+															return (byte)maxIndex;
 				case ABSOLUTE_FREQUENCY_CLASSIFICATION :  	int j;
 															for (j=0; j<classValues.length; j++)
 															{   
 																int actualValue = classValues[j];
 																if (dist[actualValue] >= probabilities[j])
 																{
-																	return actualValue;
+																	return (byte)actualValue;
 																}
 															}
 															return Instance.MISSING_VALUE;
